Promote HTTPoison wrapper to app level
